Add Migration

+++ src/main/java/as/florenko/hibernate/User.java
@@ -28,5 +28,6 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 }
+
